refactor and enable autoconfiguration for kafka
