+++ src/main/kotlin/no/novari/flyt/gateway/webinstance/InstanceMapper.kt
@@ -1,0 +1,13 @@
+package no.novari.flyt.gateway.webinstance
+
+import no.novari.flyt.gateway.webinstance.model.File
+import no.novari.flyt.gateway.webinstance.model.instance.InstanceObject
+import java.util.UUID
+
+interface InstanceMapper<T> {
+    fun map(
+        sourceApplicationId: Long,
+        incomingInstance: T,
+        persistFile: (File) -> UUID,
+    ): InstanceObject
+}
