+++ src/main/kotlin/no/novari/flyt/gateway/webinstance/exception/AbstractInstanceRejectedException.kt
@@ -1,0 +1,5 @@
+package no.novari.flyt.gateway.webinstance.exception
+
+abstract class AbstractInstanceRejectedException(
+    message: String,
+) : RuntimeException(message)
