+++ src/main/kotlin/no/novari/flyt/gateway/webinstance/exception/NoIntegrationException.kt
@@ -1,0 +1,9 @@
+package no.novari.flyt.gateway.webinstance.exception
+
+import no.novari.flyt.gateway.webinstance.model.SourceApplicationIdAndSourceApplicationIntegrationId
+
+class NoIntegrationException(
+    val sourceApplicationIdAndSourceApplicationIntegrationId: SourceApplicationIdAndSourceApplicationIntegrationId,
+) : RuntimeException(
+        "Count not find integration for ${sourceApplicationIdAndSourceApplicationIntegrationId.sourceApplicationId}",
+    )
