+++ src/main/kotlin/no/novari/flyt/gateway/webinstance/model/ArchiveCaseIdRequestParams.kt
@@ -1,0 +1,6 @@
+package no.novari.flyt.gateway.webinstance.model
+
+data class ArchiveCaseIdRequestParams(
+    val sourceApplicationId: Long,
+    val sourceApplicationInstanceId: String,
+)
