+++ src/main/kotlin/no/novari/flyt/gateway/webinstance/model/Integration.kt
@@ -1,0 +1,15 @@
+package no.novari.flyt.gateway.webinstance.model
+
+data class Integration(
+    val id: Long,
+    val sourceApplicationId: Long?,
+    val sourceApplicationIntegrationId: String?,
+    val destination: String?,
+    val state: State,
+    var activeConfigurationId: Long?,
+) {
+    enum class State {
+        ACTIVE,
+        DEACTIVATED,
+    }
+}
