+++ src/main/kotlin/no/novari/flyt/gateway/webinstance/model/SourceApplicationIdAndSourceApplicationIntegrationId.kt
@@ -1,0 +1,6 @@
+package no.novari.flyt.gateway.webinstance.model
+
+data class SourceApplicationIdAndSourceApplicationIntegrationId(
+    val sourceApplicationId: Long,
+    val sourceApplicationIntegrationId: String,
+)
