+++ src/main/kotlin/no/novari/flyt/gateway/webinstance/validation/InstanceValidationException.kt
@@ -1,0 +1,5 @@
+package no.novari.flyt.gateway.webinstance.validation
+
+class InstanceValidationException(
+    val validationErrors: List<InstanceValidationService.Error>,
+) : RuntimeException("Instance validation errors: $validationErrors")
