--- src/main/kotlin/no/novari/gateway/webinstance/exception/AbstractInstanceRejectedException.kt
@@ -1,5 +1,0 @@
-package no.novari.gateway.webinstance.exception
-
-abstract class AbstractInstanceRejectedException(
-    message: String,
-) : RuntimeException(message)
