--- src/main/kotlin/no/novari/gateway/webinstance/exception/NoIntegrationException.kt
@@ -1,9 +1,0 @@
-package no.novari.gateway.webinstance.exception
-
-import no.novari.gateway.webinstance.model.SourceApplicationIdAndSourceApplicationIntegrationId
-
-class NoIntegrationException(
-    val sourceApplicationIdAndSourceApplicationIntegrationId: SourceApplicationIdAndSourceApplicationIntegrationId,
-) : RuntimeException(
-        "Count not find integration for ${sourceApplicationIdAndSourceApplicationIntegrationId.sourceApplicationId}",
-    )
