--- src/main/kotlin/no/novari/gateway/webinstance/model/ArchiveCaseIdRequestParams.kt
@@ -1,6 +1,0 @@
-package no.novari.gateway.webinstance.model
-
-data class ArchiveCaseIdRequestParams(
-    val sourceApplicationId: Long,
-    val sourceApplicationInstanceId: String,
-)
