--- src/main/kotlin/no/novari/gateway/webinstance/model/SourceApplicationIdAndSourceApplicationIntegrationId.kt
@@ -1,6 +1,0 @@
-package no.novari.gateway.webinstance.model
-
-data class SourceApplicationIdAndSourceApplicationIntegrationId(
-    val sourceApplicationId: Long,
-    val sourceApplicationIntegrationId: String,
-)
