--- src/main/kotlin/no/novari/gateway/webinstance/model/instance/InstanceObject.kt
@@ -1,6 +1,0 @@
-package no.novari.gateway.webinstance.model.instance
-
-data class InstanceObject(
-    val valuePerKey: Map<String, String> = HashMap(),
-    val objectCollectionPerKey: MutableMap<String, Collection<InstanceObject>> = HashMap(),
-)
