--- src/main/kotlin/no/novari/gateway/webinstance/validation/InstanceValidationException.kt
@@ -1,5 +1,0 @@
-package no.novari.gateway.webinstance.validation
-
-class InstanceValidationException(
-    val validationErrors: List<InstanceValidationService.Error>,
-) : RuntimeException("Instance validation errors: $validationErrors")
