+++ src/test/kotlin/no/novari/flyt/gateway/instance/InstanceProcessorTest.kt
@@ -1,17 +1,17 @@
 package no.novari.flyt.gateway.instance
 
+import no.novari.flyt.gateway.webinstance.FileClient
+import no.novari.flyt.gateway.webinstance.InstanceMapper
+import no.novari.flyt.gateway.webinstance.InstanceProcessor
+import no.novari.flyt.gateway.webinstance.exception.FileUploadException
+import no.novari.flyt.gateway.webinstance.kafka.InstanceReceivalErrorEventProducerService
+import no.novari.flyt.gateway.webinstance.kafka.IntegrationRequestProducerService
+import no.novari.flyt.gateway.webinstance.kafka.ReceivedInstanceEventProducerService
+import no.novari.flyt.gateway.webinstance.model.File
+import no.novari.flyt.gateway.webinstance.model.Integration
+import no.novari.flyt.gateway.webinstance.model.instance.InstanceObject
+import no.novari.flyt.gateway.webinstance.validation.InstanceValidationService
 import no.novari.flyt.webresourceserver.security.client.sourceapplication.SourceApplicationAuthorizationService
-import no.novari.gateway.webinstance.FileClient
-import no.novari.gateway.webinstance.InstanceMapper
-import no.novari.gateway.webinstance.InstanceProcessor
-import no.novari.gateway.webinstance.exception.FileUploadException
-import no.novari.gateway.webinstance.kafka.InstanceReceivalErrorEventProducerService
-import no.novari.gateway.webinstance.kafka.IntegrationRequestProducerService
-import no.novari.gateway.webinstance.kafka.ReceivedInstanceEventProducerService
-import no.novari.gateway.webinstance.model.File
-import no.novari.gateway.webinstance.model.Integration
-import no.novari.gateway.webinstance.model.instance.InstanceObject
-import no.novari.gateway.webinstance.validation.InstanceValidationService
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
